+++ os/src/arch/x86_64/utils/byte_order.h
@@ -19,5 +19,5 @@ uint64_t be64_to_cpu(uint32_t be64);
 uint16_t htons(uint16_t hostshort);
 uint32_t htonl(uint32_t hostlong);
 uint16_t ntohs(uint16_t netshort);
-uint16_t ntohl(uint32_t netlong);
+uint32_t ntohl(uint32_t netlong);
 #endif
